requirements_all.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
chinesecalendar==1.8.0
requests==2.31.0
scikit-learn==1.4.0
matplotlib==3.8.2

# 爬虫依赖 (新增)
beautifulsoup4==4.12.3
lxml==5.1.0

# 核心深度学习库 (必须严格匹配版本以避免兼容性问题)
tensorflow==2.15.0
tensorflow-intel==2.15.0
keras==2.15.0
h5py==3.10.0
protobuf==4.25.3
tensorboard==2.15.1
tensorflow-estimator==2.15.0
